refactor(sdk): split errors, types, and utils into domain-specific modules
Viem-tier granularity: replace monolithic flat files with focused modules
organized by domain, each with a barrel index.ts preserving all existing
import paths.

errors/ — base, signing, encryption, transaction, credential, relayer, delegation
types/ — contract, signer, transaction, storage, credentials, callbacks
utils/ — assertions, hex, error

Co-located tests moved into each domain's __tests__/ directory.
