+++ packages/sdk/src/errors/delegation.ts
@@ -1,0 +1,38 @@
+import { ZamaError, ZamaErrorCode } from "./base";
+
+// Delegation errors — the SDK does NOT auto-map ACL contract reverts to these.
+// They are exported so dApp code can catch and re-throw them when parsing
+// on-chain revert reasons (e.g. via viem's `decodeErrorResult`).
+// See the design doc for the list of ACL revert selectors.
+
+/** Delegation cannot target self (delegate === msg.sender). */
+export class DelegationSelfNotAllowedError extends ZamaError {
+  constructor(message: string, options?: ErrorOptions) {
+    super(ZamaErrorCode.DelegationSelfNotAllowed, message, options);
+    this.name = "DelegationSelfNotAllowedError";
+  }
+}
+
+/** Only one delegate/revoke per (delegator, delegate, contract) per block. */
+export class DelegationCooldownError extends ZamaError {
+  constructor(message: string, options?: ErrorOptions) {
+    super(ZamaErrorCode.DelegationCooldown, message, options);
+    this.name = "DelegationCooldownError";
+  }
+}
+
+/** No active delegation found for this (delegator, delegate, contract) tuple. */
+export class DelegationNotFoundError extends ZamaError {
+  constructor(message: string, options?: ErrorOptions) {
+    super(ZamaErrorCode.DelegationNotFound, message, options);
+    this.name = "DelegationNotFoundError";
+  }
+}
+
+/** The delegation has expired. */
+export class DelegationExpiredError extends ZamaError {
+  constructor(message: string, options?: ErrorOptions) {
+    super(ZamaErrorCode.DelegationExpired, message, options);
+    this.name = "DelegationExpiredError";
+  }
+}
