+++ packages/sdk/src/errors/index.ts
@@ -1,0 +1,12 @@
+export { ZamaError, ZamaErrorCode, matchZamaError } from "./base";
+export { SigningRejectedError, SigningFailedError, wrapSigningError } from "./signing";
+export { EncryptionFailedError, DecryptionFailedError } from "./encryption";
+export { ApprovalFailedError, TransactionRevertedError } from "./transaction";
+export { KeypairExpiredError, InvalidKeypairError, NoCiphertextError } from "./credential";
+export { RelayerRequestFailedError, ConfigurationError } from "./relayer";
+export {
+  DelegationSelfNotAllowedError,
+  DelegationCooldownError,
+  DelegationNotFoundError,
+  DelegationExpiredError,
+} from "./delegation";
