+++ packages/sdk/src/errors/relayer.ts
@@ -1,0 +1,21 @@
+import { ZamaError, ZamaErrorCode } from "./base";
+
+/** Relayer HTTP request failed. */
+export class RelayerRequestFailedError extends ZamaError {
+  /** HTTP status code from the relayer, if available. */
+  readonly statusCode: number | undefined;
+
+  constructor(message: string, statusCode?: number, options?: ErrorOptions) {
+    super(ZamaErrorCode.RelayerRequestFailed, message, options);
+    this.name = "RelayerRequestFailedError";
+    this.statusCode = statusCode;
+  }
+}
+
+/** SDK configuration is invalid (e.g. forbidden chain ID, unsupported type). */
+export class ConfigurationError extends ZamaError {
+  constructor(message: string, options?: ErrorOptions) {
+    super(ZamaErrorCode.Configuration, message, options);
+    this.name = "ConfigurationError";
+  }
+}
