+++ packages/sdk/src/utils/index.ts
@@ -1,0 +1,9 @@
+export { toError } from "./error";
+export { prefixHex, unprefixHex } from "./hex";
+export {
+  assertObject,
+  assertString,
+  assertArray,
+  assertStringProp,
+  assertCondition,
+} from "./assertions";
